+++ HISTORY.md
@@ -6,8 +6,12 @@
 * **Notable Changes:**
     * same blocks with empty variadic inputs compare as equal regardless of their arity
 * **Notable Fixes:**
+    * removed distinction between number and string keys in "analyze" 
 * **Documentation Updates:**
 * **Translation Updates:**
+
+### 2022-01-01
+* extensions: removed distinction between number and string keys in "analyze"
 
 ### 2021-12-22
 * new dev version